External Audit Storage - Handler Refactor
* Refactor `tctl` external_audit_storage resource to use Handler pattern
