+++ tool/tctl/common/collection.go
@@ -35,7 +35,6 @@ import (
 	healthcheckconfigv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/healthcheckconfig/v1"
 	loginrulepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/loginrule/v1"
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/api/types/externalauditstorage"
 	"github.com/gravitational/teleport/api/types/label"
 	"github.com/gravitational/teleport/api/types/secreports"
 	apiutils "github.com/gravitational/teleport/api/utils"
@@ -300,39 +299,6 @@ func (c *integrationCollection) WriteText(w io.Writer, verbose bool) error {
 		})
 	}
 	headers := []string{"Name", "Type", "Spec"}
-	t := asciitable.MakeTable(headers, rows...)
-	_, err := t.AsBuffer().WriteTo(w)
-	return trace.Wrap(err)
-}
-
-type externalAuditStorageCollection struct {
-	externalAuditStorages []*externalauditstorage.ExternalAuditStorage
-}
-
-func (c *externalAuditStorageCollection) Resources() (r []types.Resource) {
-	for _, a := range c.externalAuditStorages {
-		r = append(r, a)
-	}
-	return r
-}
-
-func (c *externalAuditStorageCollection) WriteText(w io.Writer, verbose bool) error {
-	var rows [][]string
-	for _, a := range c.externalAuditStorages {
-		rows = append(rows, []string{
-			a.GetName(),
-			a.Spec.IntegrationName,
-			a.Spec.PolicyName,
-			a.Spec.Region,
-			a.Spec.SessionRecordingsURI,
-			a.Spec.AuditEventsLongTermURI,
-			a.Spec.AthenaResultsURI,
-			a.Spec.AthenaWorkgroup,
-			a.Spec.GlueDatabase,
-			a.Spec.GlueTable,
-		})
-	}
-	headers := []string{"Name", "IntegrationName", "PolicyName", "Region", "SessionRecordingsURI", "AuditEventsLongTermURI", "AthenaResultsURI", "AthenaWorkgroup", "GlueDatabase", "GlueTable"}
 	t := asciitable.MakeTable(headers, rows...)
 	_, err := t.AsBuffer().WriteTo(w)
 	return trace.Wrap(err)
